requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
fastapi==0.115.6
greenlet==3.1.1
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0